kn-py-vmdata-psql-fn/requirements.txt
cloudevents==1.11.0
Flask==2.3.2
  click==8.1.3
  importlib-metadata==4.11.4
    zipp==3.19.1
  itsdangerous==2.1.2
  Jinja2==3.1.5
    MarkupSafe==2.1.1
  Werkzeug==3.0.6
pip==25.0.1
setuptools==70.0.0
psycopg2-binary==2.9.9